feat(ui): prepare to separate login and register
