+++ app/pages/register.vue
@@ -1,4 +1,4 @@
-<script setup lang="ts">
+<script lang="ts" setup>
 	import { authClient } from '~~/lib/auth-client';
 
 	const session = authClient.useSession();
@@ -20,26 +20,18 @@
 			await navigateTo('/home');
 		}
 	};
-
-	const handleLoginSubmit = async (event: Event) => {
-		event.preventDefault();
-		const { data, error } = await authClient.signIn.email({
-			email: email.value,
-			password: password.value,
-		});
-		if (error) {
-			console.error('Error signing in:', error);
-		} else {
-			console.log('Signed in successfully:', data);
-			await navigateTo('/home');
-		}
-	};
 </script>
 
 <template>
-	<div>
+	<div
+		class="shadow-4xl inline-block flex-col items-center rounded-xl border-1 border-gray-300 p-6"
+	>
 		<h1>Register form</h1>
-		<UForm novalidate @submit="handleRegisterSubmit">
+		<UForm
+			novalidate
+			class="flex flex-col items-center gap-4"
+			@submit="handleRegisterSubmit"
+		>
 			<UInput
 				v-model="email"
 				type="email"
@@ -65,25 +57,7 @@
 			/>
 			<UButton type="submit">Register</UButton>
 		</UForm>
-
-		<h1>Login form</h1>
-		<UForm novalidate @submit="handleLoginSubmit">
-			<UInput
-				v-model="email"
-				type="email"
-				placeholder="Email"
-				required
-				autocomplete="email"
-				autofocus
-			/>
-
-			<UInput
-				v-model="password"
-				type="password"
-				placeholder="Password"
-				required
-			/>
-			<UButton type="submit">Login</UButton>
-		</UForm>
 	</div>
 </template>
+
+<style></style>
